+++ examples/simple.js
@@ -8,13 +8,16 @@ const Example = React.createClass({
   getInitialState() {
     return {
       percent: 30,
+      value: 30,
       color: '#3FC7FA',
     };
   },
   changeState() {
     const colorMap = ['#3FC7FA', '#85D262', '#FE8C6A'];
+    const value = parseInt(Math.random() * 100, 10);
     this.setState({
-      percent: parseInt(Math.random() * 100, 10),
+      percent: value,
+      value: value,
       color: colorMap[parseInt(Math.random() * 3, 10)],
     });
   },
@@ -25,6 +28,7 @@ const Example = React.createClass({
     const circleContainerStyle = {
       width: '250px',
       height: '250px',
+      display: 'inline-block',
     };
     return (
       <div>
@@ -39,6 +43,45 @@ const Example = React.createClass({
             strokeWidth="6"
             strokeLinecap="square"
             strokeColor={this.state.color}
+          />
+        </div>
+        <div style={circleContainerStyle}>
+          <Circle
+            value={this.state.value}
+            openWidth={70}
+            strokeWidth="6"
+            strokeLinecap="square"
+            strokeColor={this.state.color}
+          />
+        </div>
+        <div style={circleContainerStyle}>
+          <Circle
+            value={this.state.value}
+            openWidth={70}
+            strokeWidth="6"
+            strokeLinecap="square"
+            strokeColor={this.state.color}
+            beginPosition="bottom"
+          />
+        </div>
+        <div style={circleContainerStyle}>
+          <Circle
+            value={this.state.value}
+            openWidth={70}
+            strokeWidth="6"
+            strokeLinecap="square"
+            strokeColor={this.state.color}
+            beginPosition="left"
+          />
+        </div>
+        <div style={circleContainerStyle}>
+          <Circle
+            value={this.state.value}
+            openWidth={70}
+            strokeWidth="6"
+            strokeLinecap="square"
+            strokeColor={this.state.color}
+            beginPosition="right"
           />
         </div>
         <p>