+++ src/Circle.js
@@ -1,27 +1,61 @@
 /* eslint-disable react/prop-types */
-import React from 'react';
+import React, { PropTypes } from 'react';
 import mixin from './mixin';
 
 export default React.createClass({
+  propTypes: {
+    beginPosition: PropTypes.oneOf(['top', 'bottom', 'left', 'right']),
+  },
   mixins: [mixin],
+
   render() {
     const {
       prefixCls, strokeWidth, trailWidth, strokeColor,
       trailColor, strokeLinecap, percent, style, className,
+      value, minValue = 0, maxValue = 100, openWidth = 0, beginPosition,
       ...restProps,
     } = this.props;
 
     const radius = (50 - strokeWidth / 2);
-    const pathString = `M 50,50 m 0,-${radius}
-     a ${radius},${radius} 0 1 1 0,${2 * radius}
-     a ${radius},${radius} 0 1 1 0,-${2 * radius}`;
+    let beginPositionX = 0;
+    let beginPositionY = -radius;
+    let endPositionX = 0;
+    let endPositionY = -2 * radius;
+    switch (beginPosition) {
+      case 'left':
+        beginPositionX = -radius;
+        beginPositionY = 0;
+        endPositionX = 2 * radius;
+        endPositionY = 0;
+        break;
+      case 'right':
+        beginPositionX = radius;
+        beginPositionY = 0;
+        endPositionX = -2 * radius;
+        endPositionY = 0;
+        break;
+      case 'bottom':
+        beginPositionY = radius;
+        endPositionY = 2 * radius;
+        break;
+      default:
+    }
+    const pathString = `M 50,50 m ${beginPositionX},${beginPositionY}
+     a ${radius},${radius} 0 1 1 ${endPositionX},${-endPositionY}
+     a ${radius},${radius} 0 1 1 ${-endPositionX},${endPositionY}`;
     const len = Math.PI * 2 * radius;
-    const pathStyle = {
-      strokeDasharray: `${len}px ${len}px`,
-      strokeDashoffset: `${((100 - percent) / 100 * len)}px`,
+    const trailPathStyle = {
+      strokeDasharray: `${len - openWidth}px ${len}px`,
+      strokeDashoffset: `-${openWidth / 2}px`,
       transition: 'stroke-dashoffset 0.3s ease 0s, stroke 0.3s ease',
     };
 
+    const currentValue = value || percent;
+    const strokePathStyle = {
+      strokeDasharray: `${(currentValue - minValue) / (maxValue - minValue) * (len - openWidth)}px ${len}px`,
+      strokeDashoffset: `-${openWidth / 2}px`,
+      transition: 'stroke-dashoffset 0.3s ease 0s, stroke 0.3s ease',
+    };
     return (
       <svg
         className={`${prefixCls}-circle ${className}`}
@@ -35,6 +69,7 @@ export default React.createClass({
           stroke={trailColor}
           strokeWidth={trailWidth || strokeWidth}
           fillOpacity="0"
+          style={trailPathStyle}
         />
         <path
           className={`${prefixCls}-circle-path`}
@@ -43,8 +78,10 @@ export default React.createClass({
           stroke={strokeColor}
           strokeWidth={strokeWidth}
           fillOpacity="0"
-          ref={(path) => { this.path = path; }}
-          style={pathStyle}
+          ref={(path) => {
+            this.path = path;
+          }}
+          style={strokePathStyle}
         />
       </svg>
     );