+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ---
 
+## 2.1.0
+- Add `value` `minValue` `maxValue` `openWidth` `beginPosition` props.
+
 ## 2.0.0
 
 - refactor code